refactor: split batch workflow internals
